Revert "Move Tracker to evm-gateway"
This reverts commit 5d3ba373c43b84e05ce98d6b517db98b9d767cf1.
